+++ extensions/bitbucket/src/helpers/preferences.ts
@@ -6,4 +6,5 @@ export interface Preferences {
   workspace: string;
   email: string;
   apiToken: string;
+  maxRepoAgeDays: string;
 }